backend/requirements.txt
# Flask & Web Server
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
python-dotenv==1.0.1
Flask-Limiter==3.5.1
flasgger==0.9.7.1

# AI & Machine Learning - Windows-compatible versions
transformers==4.40.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.0+cpu
torchvision==0.18.0+cpu
torchaudio==2.3.0+cpu
sentencepiece==0.2.0

# Firebase
firebase-admin==6.5.0

# File Processing & NLP - Using lighter alternatives
python-docx==1.1.2
PyMuPDF==1.24.2
nltk==3.8.1
pdfplumber==0.11.0

# File Exporting
reportlab==4.2.0

# Utilities
requests==2.31.0
